Before stating to use Automapper

+++ RegiStar/Windows/Admin.xaml.cs
@@ -32,8 +32,8 @@ namespace RegiStar.Windows
 
         private void ddlClass_SelectionChanged(object sender, System.Windows.Controls.SelectionChangedEventArgs e)
         {
+            //Every time the user selected a new item, update the student class list.
             listStudent.ItemsSource = AdminView.getStudentsInClass();
-
         }
 
         //Add student to classlist.
@@ -47,30 +47,46 @@ namespace RegiStar.Windows
         //New student
         private void btnNewStudent_Click(object sender, RoutedEventArgs e)
         {
-            PeopleView newStudent = new PeopleView(new tblStudent());
+            PeopleView newStudent = new PeopleView(new tblStudent(), AdminView.NewestStudent);
             newStudent.Show();
+            listStudent.ItemsSource = AdminView.getStudentsInClass();
+
         }
 
         //Edit Student
         private void btnEditStudent_Click(object sender, RoutedEventArgs e)
         {
-            PeopleView editStudent = new PeopleView(AdminView.selectedStudentList);
-            editStudent.Show();
+            if(AdminView.selectedStudentList != null)
+            {
+                PeopleView editStudent = new PeopleView(AdminView.selectedStudentList);
+                editStudent.Show();
+            }
+            else
+            {
+                MessageBox.Show("You haven't selected a student from the dowpdown list yet!", "ERROR: Please make a selection.", MessageBoxButton.OK, MessageBoxImage.Warning);
+            }
+
         }
 
         //New teacher
         private void btnNewTeacher_Click(object sender, RoutedEventArgs e)
         {
-            PeopleView newTeacher = new PeopleView(new tblTeacher());
+            PeopleViewTeacher newTeacher = new PeopleViewTeacher(new tblTeacher(), AdminView.NewestTeacher);
             newTeacher.Show();
         }
 
         //Edit Teacher
         private void btnEditTeacher_Click(object sender, RoutedEventArgs e)
         {
-            //TODO: FIND A WAY TO SEND THE TEACHER INFO.
-            PeopleView editTeacher = new PeopleView(AdminView.selectedCourseList.tblTeacher);
-            editTeacher.Show();
+            if(AdminView.TeacherForCourse != null)
+            {
+                PeopleViewTeacher editTeacher = new PeopleViewTeacher(AdminView.TeacherForCourse);
+                editTeacher.Show();
+            }
+            else
+            {
+                MessageBox.Show("You haven't selected a coruse from the dowpdown list yet!", "ERROR: Please make a selection.", MessageBoxButton.OK, MessageBoxImage.Warning);
+            }
         }
 
         //New Course
@@ -83,26 +99,59 @@ namespace RegiStar.Windows
         //Edit Course
         private void EditCourse_Click(object sender, RoutedEventArgs e)
         {
-            CourseWindow editCourseWindow = new CourseWindow(AdminView.selectedCourseList);
-            editCourseWindow.Show();
+            if(AdminView.selectedCourseList != null)
+            {
+                CourseWindow editCourseWindow = new CourseWindow(AdminView.selectedCourseList);
+                
+                editCourseWindow.Show();
+            }
+            else
+            {
+                MessageBox.Show("You haven't selected a coruse from the dowpdown list yet!", "ERROR: Please make a selection.", MessageBoxButton.OK, MessageBoxImage.Warning);
+            }
         }
 
         //Remove all students
         private void Button_Click(object sender, RoutedEventArgs e)
         {
-            AdminView.DeleteAll();
-            listStudent.ItemsSource = AdminView.getStudentsInClass();
+            if(AdminView.selectedCourseList != null)
+            {
+                AdminView.DeleteAll();
+                listStudent.ItemsSource = AdminView.getStudentsInClass();
+            }
+            else
+            {
+                MessageBox.Show("You haven't selected a coruse from the dowpdown list yet!", "ERROR: Please make a selection.", MessageBoxButton.OK, MessageBoxImage.Warning);
+            }
         }
 
+        //Remove the selected coruse
         private void btnRemoveClass_Click(object sender, RoutedEventArgs e)
         {
-            AdminView.RemoveClass();
+            if(AdminView.selectedCourseList != null)
+            {
+                AdminView.RemoveClass();
+                ddlClass.ItemsSource = AdminView.getClasses();
+            }
+            else
+            {
+                MessageBox.Show("You haven't selected a coruse from the dowpdown list yet!", "ERROR: Please make a selection.", MessageBoxButton.OK, MessageBoxImage.Warning);
+            }
+            
         }
 
+        //Remove selected student
         private void btnRemoveSelected_Click(object sender, RoutedEventArgs e)
         {
-            AdminView.RemoveSelectedStudent();
-            listStudent.ItemsSource = AdminView.getStudentsInClass();
+            if(AdminView.selectedStudentsInClass != null)
+            {
+                AdminView.RemoveSelectedStudent();
+                listStudent.ItemsSource = AdminView.getStudentsInClass();
+            }
+            else
+            {
+                MessageBox.Show("You haven't selected a student from the list yet!", "ERROR: Please make a selection.", MessageBoxButton.OK, MessageBoxImage.Warning);
+            }
         }
     }
 }
